Change packaging rules to operate on strings instead of path fragments.
The upcoming path refactor will normalize all path fragments upon creation. That is fine 99% of the time, but sometimes we want to disallow non-normalized paths on rule attributes. Even this isn't usually a problem since we can validate the string prior to putting it in a path fragment.

However, in the case of the packaging rule we do not know the rules of validation until the packaging rule is *consumed* by some other rule. Therefore, retain input as a string on these rules all the way through.

We don't really do a lot of path fragmenty stuff with these strings. The main drawback is losing a bit of type safety / readability.

SKIP_KOKORO

PiperOrigin-RevId: 181760613