Create ndk wrong path in TMPDIR, not SRCDIR
test_android_ndk_repository_wrong_path presumes the wrong ndk path
'some_dir' to not exist. This test will work initially, but break if run
incrementally, because of the some_dir remaining in the TEST_SRCDIR (the
runfiles root for this test). Switching this over to use the ephemeral
TEST_TMPDIR instead.

Closes #4465.

PiperOrigin-RevId: 182083666