Unmark InterruptibleTest as flaky, it hasn't flaked in a month.
I've attempted to recreate the flakiness, to see if changing the test from a wait-based approach to actual locking would improve the flakiness, but the test seems to not have flaked since it was broken out into its own target, so it seems that might just add unnecessary complexity.

RELNOTES: None.
PiperOrigin-RevId: 183235405