Stop warning about resource filters with "deprecated" qualifiers
Resource qualifiers of the form "en_US" are deprecated for resources (in favor
of qualifiers of the form "en-rUS") but are apparently the recommended (if not
only) way of specifying locales in resource filter strings.

Removing the warning (and reverting any changes that tried following it) should
solve this problem - the original filter strings from the rule are passed
unmodified to aapt, but these filters are still "fixed" so that
android_ide_common (which doesn't support the "deprecated" qualifier form) can
still be used to filter resources in analysis.

RELNOTES: none
PiperOrigin-RevId: 183830253