http_archive: verify that unicode characters are OK in tar archives
Add a test verifying that http_archive can extract a tar archive
containing unicode characters. While such files cannot be referred
to by labels, it is still important that the archive can be extracted.
Also fix that use case on Darwin, by appropriately reencoding the string,
so that the Files java standard library can encode it back to what we
had in the first place.

Work-around for #1653, showing that http_archive from @bazel_tools can
be used; however, the issue still remains for zip archives.

Change-Id: If944203bf618c21705af676347d8591ab015d559
PiperOrigin-RevId: 183987726