sha256 python tool: stream instead of reading whole file beforehand
This is a rather small change to a Python tool used to produce a SHA256 hash.  Currently, the whole file is loaded in memory before computing the hash, which causes problem when large files are processed.  For instance, github.com/bazelbuild/rules_docker uses it to compute the hash of Docker images, which can be multiple GB in size.  This PR avoids the tool to cause issues in a limited-memory environment.

Closes #4243.

PiperOrigin-RevId: 184518900