BEP: ensure that TestAttempts are reported after their target
Ensure that each test attempt is only reported after the report of the
completion of the build of the corresponding test target. Normally this
should happen anyway, but due to races on the internal event bus, the
order of the report might be messed up. So add an explicit order constraint.

Change-Id: I4d325bc31a46dcdf8763ba5416b5135a0978536e
PiperOrigin-RevId: 184825306