Automated rollback of commit 873f343fefbb08048e7e75c482843b9e68954de6.
*** Reason for rollback ***

Breaking targets in 2018.02.08 and 2018.02.09 nightly TGPs

There were some merge conflicts in this rollback CL. See unknown commit for the unresolved version. I kept the YOURS sections, and ran fiximports.py to remove unused imports.

See b/73157879

*** Original change description ***

C++: Remove last instatiation of CppModel outside CcLibraryHelper.

This is the second try for this CL. The first one caused Blaze to crash when building Exoblaze as shown in b/72936965. In this CL I fix the condition of when to generate non-PIC compilation actions for WrapCcHelper.

RELNOTES:none
PiperOrigin-RevId: 185203398