Let ijar handle jars with incorrect class files
Some jars in Maven Central (e.g. aspectjtools-1.8.13.jar) contain
invalid class files. When such a jar is used in a Bazel project the ijar
utility returns with an error, which makes it impossible to use the jar.

By continuing on an invalid class file after emitting an error message
those jars can be used in a Bazel project.