Fix screen flicker caused by small writes.
ExperimentalEventHandler is in charge of composing the pretty progress bar
that Bazel displays.  This progress bar is a multi-line message with control
characters printed on the terminal.

The progress bar was composed by issuing many individual writes to an
AnsiTerminal.  Because the AnsiTerminal in this case was backed by an error
stream (which are unbuffered), each of these writes resulted in a gRPC to
the Bazel client to write the message to the console.  gRPC calls are much
more expensive than calls to a file descriptor, and, in general, even small
writes to a file descriptor should be avoided when preparing long messages.

To fix this, fully buffer the output messages sent to the AnsiTerminal until
explicitly flushed.  ExperimentalEventHandler was already doing the right
thing regarding flushes but did not account for the fact that each write
would be (unintentionally) sent directly to the terminal.

The flicker was significant: on a pathological case (building sandboxfs with
Bazel on my MacBook Pro 13" on macOS), this change shaves about 5 seconds of
build time on the previous 45 second-long build.  I think this only happened
with "bazel run" and "bazel test" invocations and not "bazel build", but
I haven't really confirmed this.

RELNOTES: None.
PiperOrigin-RevId: 185405892