Add the BUILD_WORKSPACE_DIRECTORY and BUILD_WORKING_DIRECTORY variables to "blaze run --direct_run" so that the called binary knows about the working directory the client was called from.
Its cwd is its runfiles directory and if not for the fact that we have to convey *two* directories to it, I'd have considered changing that. As it is, however, we can't convey two directories with the cwd of the binary so we have to use environment variables.

RELNOTES[NEW]: "blaze run --direct_run" now exports the BUILD_{WORKSPACE,WORKING}_DIRECTORY variables to tell the binary about the cwd of the client and the workspace root.

PiperOrigin-RevId: 185515884