[Skylint] Do not crash ControlFlowChecker on nested functions.
ControlFlowChecker has a precondition check that assumes nested functions do not occur. While this assumption is reasonable for a valid Skylark syntax, linter can actually be invoked on malformed files and users would get a stack trace instead of a human-readable linter error.

Alternative and possibly a better strategy would to not run CFChecker in case parse errors are detected.
fixes #4511

Closes #4512.

PiperOrigin-RevId: 185538897