Implement the config(expr, word) cquery function.
expr - the expression to be evaluated
word - the configuration (represented by the strings 'host', 'target', or 'null') to try to find the result(s) of 'expr' in. If some but not all results of expr can be found in the specified config, then the subset that can be is returned. If no results of expr can be found in the specified config, then an error is thrown.

PiperOrigin-RevId: 185572590