Skipping writing FILTERED tests to test.xml
This fixes #3201 by preventing tests that haven't actually run to be written to the test.xml. This is consistent with how e.g. surefire reports work, tests that were filtered out do not appear in the xml.
This allows changing the Bazel plugin in such a way that does not depend on `time` being 0.0.

Closes #4596.

PiperOrigin-RevId: 185810400