Provide a hacky way to get an ActionInput from BinTools
We have two places in internal code that require an ActionInput for an embedded
tool. The embedded tool is currently not declared as an input to the
corresponding actions, which in turn causes the remote execution to stat the
file directly through SingleBuildFileCache. I'm working on removing
SingleBuildFileCache, so this needs to be solved. This is the quick-and-dirty
fix. A more principled fix might be to check in the relevant tools instead of
embedding them into Bazel, but it's also significantly more work.

PiperOrigin-RevId: 186292203