Decprecate native http_archive
Since the skylark implementation of http_archive provides all
the features available in the native http_archive and a few
more (patching), there is no need to have to rule implementations
doing the same thing. Hence deprecate the native one as part
of our plan of moving more functionality to skylark.

RELNOTES: The native http_archive rule is deprecated. Use the
  Skylark version available via load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
  instead.

Change-Id: I107c2f25f5a37c67f56b4362e7c9d9ade8428c16
PiperOrigin-RevId: 186314624