Fix test name in fallback xml when --run_under is used.
When --run_under is used, the run-under wrapper will be injected in front of the normal test command line. So, it's incorrect to use the beginning of the command line to infer the test name. However, we can always get the underlying test executable with the TEST_BINARY environmental variable.

Rename TEST_NAME to EXE to avoid future confusion.

Fixes https://github.com/bazelbuild/bazel/issues/4588.

Change-Id: I6fd05cca5e5441c13ee16290c4028ec84adec983
PiperOrigin-RevId: 186590311