Recompile linkstamps whenever any input of the owning action changes.
Linkstamps were not re-built when only volatile data changed, i.e. when we
modified cc_binary source, linkstamp was not recompiled so we got old
timestamps. The proper behavior is to recompile linkstamp whenever any input to
cc_binary linking action changes. And the implementation in this cl solves this
by adding all linking inputs as inputs to linkstamp compile action.

RELNOTES: None.
PiperOrigin-RevId: 186595143