windows,tests: fix //src/test/py/bazel:py_test
The test was failing because it assumed Bazel was
generating a runfiles tree for py_binary.

On Windows, Bazel actually builds a zip file which
contains the py_binary's runfiles tree.

This commit adjusts the tests to assert the zip
file's contents when the test is running on
Windows.

Fixes https://github.com/bazelbuild/bazel/issues/4673

Change-Id: I7544bff6c1866249b8e7982b1a106202d78e10d9
PiperOrigin-RevId: 186770549