Fixes skyframe evaluation to correctly handle interrupts during #doMutatingEvaluation
When multiple keys need to be evaluated, we may schedule keys before informing
progress on done nodes, which can throw InterruptExceptions. If the main thread
is interrupted during #informProgressReceiverThatValueIsDone, then we may not
properly track evaluations which are already scheduled. It could cause threads
continue to run after the main loop is closed.

RELNOTES: None
PiperOrigin-RevId: 186801930