Keep the order of select branches the same as it was in the BUILD file.
This is a change from behavior introduced in unknown commit, which was added to preserve determinism even though the ordering of dicts in Python is non-deterministic. Now that we don't call into Python anymore, this is not necessary.

RELNOTES: None (select branches are mutually exclusive, so this is a no-op)
PiperOrigin-RevId: 187022575