Optimize SkylarkNestedSet construction
Only check/update type if the types of elements that we're adding changes. In
particular, the "is DICT/LIST" check can get expensive when run on every single
inserted element.

According to a hacky unit test (create 10M element SkylarkNestedSet 10x over) this is ~60%
faster (~18s -> ~7s).

This change is intended to be a quick fix, there's surely a more principled way
to make skylark's type checking more efficient...

PiperOrigin-RevId: 187062547