Java: Stops adding C++ providers directly.
unknown commit fixes the only related broken target showing up in a TGP after
this change. It seems that this provider was only ever used through
java_wrap_cc with java_library dependencies which themselves had cc_library
dependencies, i.e. java_wrap_cc -> java_library -> cc_library. This dependency
is not needed, the cc_library can be added directly to java_wrap_cc to get the
same behavior and this is the way it should be done instead of relying on
transitive dependencies.

The reason for making this change right now is that with the Skylark C++ API, I
will be adding attr("deps").mandatoryProviders(CcCompilationInfo,
CcLinkParamsInfo) to C++ rules so that they can take any Skylark rule and not
having to explicitly hard-code the name of those Skylark rules in Blaze.
However, Java meets this criteria and the change would allow C++ rules to
depend on Java. This is not something that was supported before and is not
something we plan to support.

RELNOTES:none
PiperOrigin-RevId: 187158176