Optimize Path (and the underlying String) allocations in UnixGlob for the case where a pattern like '*.blah' doesn't match all the dirents.
Before the recent Path.java rewrite, the unconditional allocation here was free/amortized (in the sense that if Blaze ever had to consider the path in question, then we'd be doing the Path allocation anyway and caching the object). But now it's potentially GC churny, especially in the case where a single BUILD file has multiple glob expressions.

RELNOTES: None
PiperOrigin-RevId: 187185573