Make Location#printLocation private
This is only used by the static method of the same name, which is only used by RuleFactory.
This method is showing up in profiles as creating more garbage than I'm happy with, but not
enough to make me drop everything and fix it now. Instead make it private so folks are less
inclined to use it later...

PiperOrigin-RevId: 187208467