ci,windows: fix "bazel-rules-tests"
Fix the failing test in
//src/test/java/com/google/devtools/build/lib:bazel-rules-tests
on Windows.

The problem was that the test setup read a
resource file and assumed line endings to be "\n",
but on Windows this resource file was generated
with "\r\n" endings.

The test setup code then put those strings into a
set.

Finally, the test tried to match a string in this
set, but because the entries had an invisible "\r"
at the end, the matching failed.

This commit trims the entries before putting them
in the set.

Fixes https://github.com/bazelbuild/bazel/issues/4752

Closes #4754.

PiperOrigin-RevId: 187608803