Don't use computeIfAbsent in action graph dumping.
computeIfAbsent may throw (on best effort basis) a
ConcurrentModificationException if you change the underlying Map in the
computeIfAbsent call, for example when you call computeIfAbsent recursively in
the mappingFunction.
NestedSets are an inherent recursive structure, so we need to call
computeIfAbsent recursively in the action graph dump.

Since computeIfAbsent doesn't throw the exception all the time, it's hard to
come up with a reliable test case.

RELNOTES: None
PiperOrigin-RevId: 188151283