Fixed missing C standard library functions with NDK r15 and r16
Certain C++ stdlib wrapper headers are unable to find their C counterparts (e.g. math.h for cmath) that are in %ndk%/sysroot/usr/include. This is because the -isystem for the include path was added with addCompilerFlag instead of addUnfilteredCxxFlag. The former is filtered out when compiling C code, whereas the latter keeps it.

Fixes https://github.com/bazelbuild/bazel/issues/4777

RELNOTES: None.
PiperOrigin-RevId: 188241254