Add an interface to interact with sandboxfs.
The new SandboxfsProcess interface allows interacting with sandboxfs.
There are two implementations: RealSandboxfsProcess, which spawns the
sandboxfs binary, and FakeSandboxfsProcess, which mimics what sandboxfs
does but using symlinks and is intended for testing purposes only.

The RealSandboxfsProcess implementation works but still carries many
TODOs.  The most "painful" one may be that the test requires manual
invocation because we do not yet have an easy way to integrate with
sandboxfs.  That will be solved later on; for now this is sufficient
for initial testing.

RELNOTES: None.
PiperOrigin-RevId: 188347393