Automated rollback of commit 3edf41b70de9bb1a8702d0342beeb2ad13898d71.
*** Reason for rollback ***

Rollback was requested by original authors @hmemcpy and @ittaiz in #3201:

"We found a problem with this patch... seems that tests that are added dynamically by the test runner (in our case, specs2 'examples' that are generated with Fragments.foreach) do not appear in the xml!"

This should be part of 0.12.0-rc1, otherwise that release will have the above mentioned regression.

*** Original change description ***

Skipping writing FILTERED tests to test.xml

This fixes #3201 by preventing tests that haven't actually run to be written to the test.xml. This is consistent with how e.g. surefire reports work, tests that were filtered out do not appear in the xml.
This allows changing the Bazel plugin in such a way that does not depend on `time` being 0.0.

Closes #4596.

PiperOrigin-RevId: 188455315