Discontinue building jformatstring from source
This change stops building jformatstring from the source and consumes
the artifacts from Central. As the consequence, the sources are removed
from the Bazel tree and we discontinue packaging the sources in Bazel
source artifact.

GPL v2 with Classpath exception sources have to be still included in
third_party:gpl-srcs rule. The source jar is included in this rule, that
is used in embeded_tools archive.