Integrate flatten_virtual_headers
When flatten virtual headers is set, we add all headers to the
headermap.

This allows a user to include a Header, "path/to/header" as "header.h"

Together with include_prefix, this allows bazel to support Xcode style
includes out of the box.

As a followup, we'll want to make this work without headermaps.

+ misc cleanups from rebasing