Allow genrule dependencies in cc_ build rules.
This is going to be contentious. On one hand, power and expressiveness, and on the other, non-hermetic builds.

I have some rules that need to depend on linker scripts (which are allowed dependencies), but those scripts are generated by a genrule that includes some symbols or not depending on config_settings interpreted through select expressions.

Will this ever be allowed? If not, can we get some kind of provider for a Skylark rule to apply as some kind of "yes, I accept the risks and responsibilities of being a cc_ rule dependency."?